% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_is_encrypted.R
\name{config_is_encrypted}
\alias{config_is_encrypted}
\title{Check if a config file is encrypted}
\usage{
config_is_encrypted(
  file = "conf/config.yml",
  password_names = "dbpassword",
  secret = Sys.getenv("SHINTO_PASS_SECRET")
)
}
\arguments{
\item{file}{Path to config file (no default ued)}

\item{password_names}{Field names in YAML with passwords (can be vector)}

\item{secret}{The encryption secret}
}
\description{
Checks whether a \code{config.yml} file is completely or partially encrypted.
Returns TRUE when all \code{dbpassword} (or other \code{password_names}) entries are already encrypted,
FALSE when one or more are not yet encrypted (with a message)
}
\examples{
\dontrun{
Sys.setenv(SHINTO_PASS_SECRET = "somesecret")
config_is_encrypted("conf/config.yml")
}
}
\seealso{
\code{\link{encrypt_config_file}}
}
