% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connecties.R
\name{connect}
\alias{connect}
\title{Make a Shinto Database Connection}
\usage{
connect(
  what,
  file = getOption("shintodb_config_file", "conf/config.yml"),
  pool = FALSE,
  config_entry = NULL
)
}
\arguments{
\item{what}{Which database connection in the config file? (Name of the block in the YAML)}

\item{file}{Path to the config.yml file, or set global option \verb{options(shintodb_config_file = /path/to/file)}}

\item{pool}{Logical. When TRUE, uses \code{\link[pool:dbPool]{pool::dbPool()}}, otherwise \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\description{
Make a connection with RPostgres to one of the Shinto Labs Postgres databases.
}
