% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_config_file.R
\name{decrypt_config_file}
\alias{decrypt_config_file}
\title{Decrypt a config file}
\usage{
decrypt_config_file(
  file,
  file_out,
  secret = Sys.getenv("SHINTO_PASS_SECRET"),
  password_names = "dbpassword"
)
}
\arguments{
\item{file}{Path to config.yml file}

\item{file_out}{Path to output file (can be same as \code{file})}

\item{password_names}{Names (can be vector) of fields to be encrypted. Shinto default is \code{dbpassword}}
}
\description{
This function is normally not needed by the user, as \code{\link[=connect]{connect()}} decrypts
automatically on connecting. It is used when deploying an app to rsconnect, by \code{shintoshiny::make_deploy_project}
}
