% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_decrypt.R
\name{encrypt}
\alias{encrypt}
\alias{decrypt}
\alias{string_is_encrypted}
\title{Encrypt a password}
\usage{
encrypt(x, secret = Sys.getenv("SHINTO_PASS_SECRET"))

decrypt(x, secret = Sys.getenv("SHINTO_PASS_SECRET"))

string_is_encrypted(string, secret = Sys.getenv("SHINTO_PASS_SECRET"))
}
\arguments{
\item{x}{A character string or character vector.}

\item{secret}{The secret used for decrypting/encrypting}
}
\description{
Symmetric encryption/decryption with a secret key.
Uses the \code{libsodium} encryption library (in the \code{sodium} package).
Unlike \code{safer::encrypt_string} on which this function is based, it takes a
vector argument.
}
\details{
Set the environment variable SHINTO_PASS_SECRET. This is user-specific.
Do not reuse. Otherwise, provide a \code{secret} as an
}
