% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{has_config_entry}
\alias{has_config_entry}
\title{Check if a config entry is available}
\usage{
has_config_entry(name, where, conf = NULL, config_file = "conf/config.yml")
}
\arguments{
\item{name}{The config entry, for example "data_bag"}

\item{where}{The block to look for the entry (development, production)}

\item{conf}{Optional: a config list read with \code{\link[=read_config]{read_config()}}}

\item{config_file}{Path to the config file}
}
\value{
TRUE if a config entry is found
}
\description{
Check if a config entry is available
}
