% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{read_config}
\alias{read_config}
\title{Read a config file (with read_yaml)}
\usage{
read_config(file = "conf/config.yml")
}
\arguments{
\item{file}{Typically conf/config.yml}
}
\description{
Read a config file (with read_yaml)
}
