% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintodb-package.R
\docType{package}
\name{shintodb-package}
\alias{shintodb}
\alias{shintodb-package}
\title{shintodb: Tools for Shinto DB Connections, Config Management}
\description{
Everything to work with PostgresDB in Shinto apps. Manage, encrypt, and decrypt config files. Database R6 layer with generic methods. Manage database connections automatically.
}
\keyword{internal}
